First go at async svg DAG rendering
on the new execution-details page
on the internal DAG-engines WebConsole.

Rendered DAG is with draggable tasks.
To that end, we added a "to_nodes_list" method to the DAG class
and we now store that nodes-list to db for each new execution.

Still a work-in-progress.
  - Need to customize styling for parallel/merging tasks
    as well as task-groups, among other things.
  - Plans are to :
      - allow for the use of the 'ui_css' attribute
        for tasks and task-groups
      - make this svg DAG rendering zoomable
      - add events to nodes
        (mouseover, tooltip, contextual menu, etc).
      - and, at some probably much later point,
        to memorize user moves with cookies
        (making this safely ain't gonna be easy).

Also, we made some tiny progress here
on the executions-list ui_css front, following 03dc794.
Still also a work-in-progress, itself.