Created TaskType class
in the internal DAG-engine's model.

to serve as a generic metadata holder for
types of tasks in a DAG.
In other words, a TaskType is the definition of a task
as declared (it holds properties such as task name,
ui_css, docstring, basically all that can be rendered in a DAG)
but does not hold runtime (execution-related) info.

For a given DAG execution, there can be several instances
of a given TaskType, in cases where such a TaskType
is referenced in a parallel line. A Task
(an instance of a TaskType) holds runtime-related info
(starting with id, rank, start and end timestamp, etc.)

The intent indeed is to isolate DAG-related info
from runtime-related info.

Indeed, reworked the db to include a new dedicated table,
and got rid of the newly inserted execution.nodes_list column.

Plans are to create a TaskGroup class as well
(still for the sole purpose of DAG-rendering),
with ui_css and docstring, for starters.

Also, as part of the herein commit, started bringing
the legacy 'pipeline_card' look to the execution-details page
of the WebConsole.