Finalized execution-details page header
for the internal DAG engine's WebConsole.

For the real-time part, this includes the below counters :
  - execution number, total count, ongoing count, failed count

Also:
  - moved the 'api' part to its own subdir

Most importantly, switched to multiplexed SSE connections.
1 single SSE connection handling all events on each page.
This is to get the most out of the constraint imposed
by web-browsers on number of concurrent connections
per domain over http 1.1.

Now, one can have up to 5 tabs open on
the retrain-pipelines' WebConsole over http 1.1
with no need to upgrade to http 2.

(http 2 in practice requires https, which implies certificates
signed by a trusted authority and... too cumbersome for a solution
which aims at being easily installable locally
via simple 'pip/uv install' command).

To go with that, we also implemented a tracking mechanism
so user (in http) is prompted to close some tabs
when opening more than the amount web-browsers allow
for domain concurrent (SSE) connections.
User is also informed on how to get rid of that restriction
in prod env.