Implemented tasks instances collection logic
for the Gantt diagram of the execution-details page
of the internal DAG engine's WebConsole

and started developing the 'exec_current_progress'
async endpoint, which will be responsible for
Gantt diagram at component load time.
Comes together with the 'draw_chart' method,
the drafting has also been initiated.

Incidentally,
  - changed the DAO below 2 methods
    (both related to DAG rendering)
    so they now return lists of
    model object instances and
    not serializable dicts anymore :
      - get_execution_tasktypes_list
      - get_execution_taskgroups_list
  - created the TaskExt class, to allow
    for task & tasktype attributes
    to be handled conjointly.
  - renamed 'get_execution_elements_lists'
    into 'get_execution_dag_elements_lists'
    so Gantt-related and DAG-related methods
    have clear, distinct names.

remains to :
  - address the wrong ordering of parallel tasks,
    need to sort by (potentially nested) rank.