bug-fix - collapsible grouped table - top-level style
intended to be used for the Gantt diagram component
of the execution-details page on the internal DAG engine's
WebConsole.

fixed the problem which visually translated
into programmatically inserted rows
picking up the wrong CSS style.

Also fixed a bug discovered along the way
in calling the "incrementIndexes" method
in the wrong order (which resulted in
broken paths in groups).

Also, dropped the "defaultBottomPadding" constant,
which turned out to be unnecessary.

One known open point remains :
  - there's an edge case where 1 depth of bottom bars'
    behavior seems off.