Gantt chart integration of collapsible grouped table
in the execution-details page on the internal DAG engine's
WebConsole.

Got rid of initial FastHTML table building logic
and switched to js tableData buildup for an execution
list of tasks/taskgroups/parallel-lines.
It's now fed to page load async event and
an instance of collapsible grouped table.

Incidentally, made the js more flexible
with regards to rows items.
id, name, children and style are mandatory attributes
to row items. Any additional is treated as a column
(Note : in the order they're declared, not by name).

still a lot to do :
  - adapt styling, add handling of defaults
  - add callback on group-header click
    (for updates in addition to collapsing/expanding
     rows)
  - probably others that don't come to mind yet..
  - and of course the hard parts:
      - actual progress timelines
        (which we didn't gotten to quite yet)
      - the gantt-related events streaming
        & handling
yep, a mouthful of commits to come for this.