collapsible grouped table - cells dataset attrs
primarily intended to be used for the
Gantt diagram component of the
execution-details page on the
internal DAG engine's WebConsole.

Modified the structure of the object passed on
to the init method so as to be able to
set custom dataset attributes to individual cells.
We'll use this to set start_timestamp
and end_timestamp to timeline cells of the Gantt chart.

Also,
  - wrote a first quick version
    of the "savePageAsSingleFile" method.
    More on this soon (or not, not very sure yet
    if we'll stick with it or drop it).