Gantt chart - formatted labels
in the execution-details page on
the internal DAG engine's WebConsole.

Not all cases are implemented yet, but at this point :
  - split lines in distributed sub-pipelines
  - merge task in distributed sub-pipelines
both get labels with similar (trapezoidal)
shapes and formatting in the collapsible-grouped-table
as they do in the DAG rendering.
This is to easily identify which is which
in both components in that page.

Remark : still didn't really make a decision
on how to handle default formatting quite yet.