Gantt chart - handling of default styling
in the execution-details page on
the internal DAG engine's WebConsole.

Work in progress. The mechanism has been decided upon.
For now, however, this handling of default styling
in the collapsible-grouped-table
has only been applied to distributed sub-DAGs.

Also, completing 47bf029, all tasks now have a
formatted label in that component.