Gantt chart - live new-task event streaming
in the execution-details page on
the internal DAG engine's WebConsole.

Implemented the OpenAPI schema for
the new-Task endpoint.

Upgraded the TaskExt class constructor
so it accepts a dict as single argument.

Mostly, started working on the js
gantt-events.ganttInsert function,
which builds up on the
collapsible-grouped-table.inserAt fucntion,
plus custom formatting (among other things,
in order to handle a timeline column,
shaped labels, etc..).

Work-in-progress.

At this stage :
  - all tasks are appended to table
    as standard (standalone) tasks
  - the entire flow from pipeline execution
    (via the internal DAG-engine) down to
    the execution-details webpage live update
    is pretty much done now
  - also, formatting and handling of defaults
    is dealt with too

Remains to :
    - when appropriate, embark taskgroup metadata
      into the streamed payload
    - insert tasks at the right row/group index
      based on available payload attributes
      (taskgroup_uuid, rank, and start_timestamp)
shall be rather straightforward